:sparkles: Move route info generator to its own file.
Update homepage to show routing information.
